consolidate GitHub-API wrapper modules + fix over-broad 403 rate-limit check (#2278)
Collapse the overlapping .github/scripts GitHub-API helper modules into one
retry module, route every rate-limit/transient predicate through a single
centralized classifier, and unify the three competing API cache factories.

- error_classifier.js: add exported isRateLimitError(error) — true ONLY for
  429, or 403 + (rate-limit message OR exhausted x-ratelimit-remaining header).
  A 403 that is NOT a rate limit (e.g. "Resource not accessible by
  integration") now returns false, fixing the latent over-broad bare-403
  predicate (formerly keepalive_gate.js).
- github-api-with-retry.js: absorb github_api_retry.js (withGithubApiRetry/
  computeRetryDelayMs/resolveMaxRetries/calculateBackoffDelay) and
  api-helpers.js (paginateWithBackoff/withBackoff/checkRateLimitStatus/
  createRateLimitAwareClient/extractRateLimitReset). github-rate-limited-wrapper
  is lazily required inside the paginate/check functions to break the
  require cycle. Delete github_api_retry.js and api-helpers.js.
- Route local isRateLimitError (keepalive_gate dead copy removed; comment-dedupe,
  detect-changes, agents_pr_meta_update_body) and isTransientError
  (agents_pr_meta_orchestrator, agents_pr_meta_keepalive) through error_classifier.
  agents_pr_meta_update_body keeps its RateLimitError-instance sentinel.
- github-api-cache-client.js: single getGithubApiCache (sentinel __githubApiCache);
  keepalive_gate, keepalive_loop, agents_pr_meta_keepalive drop their local copies.
- Tests: retarget api-helpers/github-api-retry suites onto the consolidated
  module; add named 403-not-rate-limit assertion (+ deliberate-break verified);
  update keepalive-loop cache test to the new sentinel.
- Mirror all changes to templates/consumer-repo and drop the two deleted
  modules from sync-manifest.yml (validate_template_sync.py green).

Closes #2278
