+++ .github/sync-manifest.yml
@@ -396,17 +396,11 @@ scripts:
   - source: .github/scripts/keepalive_prompt_composer.js
     description: "Composes prompts for keepalive Codex calls"
 
-  - source: .github/scripts/api-helpers.js
-    description: "GitHub API helper utilities"
-
   - source: .github/scripts/autopilot_metrics.js
     description: "AutoPilot metrics collection and reporting"
 
   - source: .github/scripts/auto_pilot_transitions.js
     description: "Auto-pilot transition contract and next-step helper"
-
-  - source: .github/scripts/github_api_retry.js
-    description: "Retry logic for GitHub API calls"
 
   - source: .github/scripts/github-api-with-retry.js
     description: "GitHub API retry wrapper with exponential backoff and pagination support - required by agents-auto-pilot.yml"
