+++ templates/consumer-repo/.github/scripts/keepalive_gate.js
@@ -1,7 +1,7 @@
 'use strict';
 
-const { paginateWithBackoff, withBackoff } = require('./api-helpers.js');
-const { createGithubApiCache } = require('./github-api-cache-client');
+const { paginateWithBackoff, withBackoff } = require('./github-api-with-retry.js');
+const { getGithubApiCache } = require('./github-api-cache-client');
 const { ensureRateLimitWrapped } = require('./github-rate-limited-wrapper.js');
 
 const KEEPALIVE_LABEL = 'agents:keepalive';
@@ -37,22 +37,8 @@ const RECENT_COMPLETED_LOOKBACK_SECONDS = 300; // 5 minutes
 const RATE_LIMIT_MAX_RETRIES = 3;
 const RATE_LIMIT_BASE_DELAY_MS = 2000;
 
-function getGithubApiCache({ github, core }) {
-  if (!github) {
-    return createGithubApiCache({ core });
-  }
-  if (github.__keepaliveGateApiCache) {
-    return github.__keepaliveGateApiCache;
-  }
-  const cache = createGithubApiCache({ core });
-  Object.defineProperty(github, '__keepaliveGateApiCache', {
-    value: cache,
-    enumerable: false,
-    configurable: false,
-    writable: false,
-  });
-  return cache;
-}
+// getGithubApiCache is now the single shared factory from
+// github-api-cache-client.js (sentinel __githubApiCache).
 
 async function fetchPullRequestCached({ github, owner, repo, prNumber, core }) {
   if (!github?.rest?.pulls?.get || !owner || !repo) {
@@ -91,23 +77,6 @@ async function fetchPullRequestCached({ github, owner, repo, prNumber, core }) {
  */
 function sleep(ms) {
   return new Promise((resolve) => setTimeout(resolve, ms));
-}
-
-/**
- * Check if an error is a GitHub rate limit error.
- * @param {Error|unknown} error
- * @returns {boolean}
- */
-function isRateLimitError(error) {
-  if (!error) return false;
-  const message = error instanceof Error ? error.message : String(error);
-  const status = error?.status || error?.response?.status;
-  return (
-    status === 403 ||
-    status === 429 ||
-    /rate\s*limit/i.test(message) ||
-    /secondary\s*rate\s*limit/i.test(message)
-  );
 }
 
 function toInteger(value) {
